Rename `nightly` to `feedback`
